requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==3.1.3
SQLAlchemy==2.0.21
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
selenium==4.15.0
locust==2.17.0
email-validator==2.0.0
bcrypt==4.0.1
python-dotenv==1.0.0
Pillow==10.0.1
